+++ test/select.spec.js
@@ -1,11 +1,11 @@
 import test from 'tape'
 
-import { onChange, valid } from '../src/actions'
+import { invalid, onChange, valid } from '../src/actions'
 import { getState } from '../src/select'
 import reducer from '../src/reducer'
 
 import { emptyGetStateResult } from './mock'
-import { isRequired } from './validate'
+import { invalidDomain, isRequired } from './validate'
 
 test('getState', t => {
   const props = { validate: isRequired }
@@ -22,15 +22,22 @@ test('getState', t => {
   })
   t.deepEqual(getState({ form: state }, props), dirty, 'empty value')
   let state2 = reducer(state, onChange(null, 'kai@foo.x'))
-  state2 = reducer(state2, valid(null, { key: 'kai@foo.x', value: { message: 'invalid' } }))
+  state2 = reducer(state2, valid(null, { key: 'kai@foo.x', value: { message: 'welcome' } }))
   const dirty2 = dirty.merge({
     errorMessage: null,
     hasError: false,
     isValid: true,
     status: 'success',
-    valid: { message: 'invalid' },
+    valid: { message: 'welcome' },
     value: 'kai@foo.x',
   })
   t.deepEqual(getState({ form: state2 }, {}), dirty2, 'valid is set')
+  state2 = reducer(state2, onChange(null, 'kai@foo.xx'))
+  state2 = reducer(state2,
+    invalid(null, { key: 'kai@foo.xx', value: { message: 'wrong domain' } })
+  )
+  t.equal(getState({ form: state2 }, {}).errorMessage, 'wrong domain')
+  props.validate = invalidDomain
+  t.equal(getState({ form: state2 }, props).errorMessage, 'invalid domain')
   t.end()
 })