+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## [2.0.0]
-> 2016-02-29
+> 2016-03-01
 
 * Actions are now split into groups. `fieldEvent` contains all actions related to field state. `formEvent` are form and focus events like `onChange`. `formHandler` a clone of `formEvent` where the keys replace `on` with `handle` like `handleBlur` instead of `onBlur`.
 * First arg to actions is a `prefix` array. Second arg is the `payload`.
 * Removed concept of `formId` and `fieldId` from everything in favor of a singular `prefix` array prop.
-* Added `valid` and `invalid` actions and related state keys. Should be used to store async results to values.
+* Added `valid` and `invalid` actions and related state keys. Should be used to store async results to values. The `getState` function will look in invalid for a key matching `value`. If it finds one it will use that as error.
 * Reducer is now exported as `fieldReducer` instead of default.
 * The `validate` props function is only called when `dirty`.
 * The `submit` action now expects the value to be passed as its payload. Same as `onChange`.